moved stuff to toolbox/cuckoo
